Added Exception Handling

+++ src/main/java/com/task/intro/repository/ProductRepository.java
@@ -1,7 +1,7 @@
 package com.task.intro.repository;
 
-import com.task.intro.model.ProductEntity;
+import com.task.intro.model.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository   extends JpaRepository<ProductEntity, Integer>  {
+public interface ProductRepository   extends JpaRepository<ProductEntity, Long>  {
 }
